refactor: decompose architecture — hooks, components, memory fixes
- Fix memory leak in PatientDashboard: replace forEach(async) +
  dynamic import with Promise.allSettled + cancelled cleanup flag
- Decompose app/page.tsx: extract useDashboardData hook
- Decompose SpecialistDashboard (~530→~160 lines): extract
  useSpecialistNotifications hook and PatientList component
- Fix stale closure in patient-wounds: make loadWounds accept
  patientId as explicit parameter
